split fast and ancova
