+++ UncertaintyCat.py
@@ -17,7 +17,9 @@ from modules.correlation_analysis import correlation_analysis
 from modules.hsic_analysis import hsic_analysis
 from modules.ml_analysis import ml_analysis
 from modules.morris_analysis import morris_analysis, dimensionality_reduction_page
-from modules.fast_ancova_analysis import fast_analysis, ancova_analysis
+from modules.fast_analysis import fast_analysis
+from modules.ancova_analysis import ancova_analysis
+from modules.pce_analysis import pce_analysis
 
 # Import utils
 from utils.core_utils import (
@@ -55,7 +57,7 @@ def load_model_code(selected_model_name: str) -> str:
 
 # Page configuration
 st.set_page_config(
-    page_title="UncertaintyCat | Enterprise Edition",
+    page_title="UncertaintyCat | Version 5.12",
     page_icon="🐱",
     layout="wide",
     initial_sidebar_state="expanded",
@@ -69,7 +71,7 @@ col1, col2 = st.columns([1, 5])
 with col1:
     st.image("logo.jpg", width=100)
 with col2:
-    st.markdown('<h1 class="main-header">UncertaintyCat | Version 5.11</h1>', unsafe_allow_html=True)
+    st.markdown('<h1 class="main-header">UncertaintyCat | Version 5.12</h1>', unsafe_allow_html=True)
     st.markdown('<p style="color: #7F8C8D;">Advanced Uncertainty Quantification and Sensitivity Analysis Platform</p>', unsafe_allow_html=True)
 
 # Sidebar styling and navigation
